+++ packages/components-angular/tests/e2e/specs/input-password-form.e2e.ts
@@ -14,7 +14,7 @@ const setValue = async (page) => {
 
 test.describe('Angular forms integration', () => {
   test.beforeEach(async ({ page }) => {
-    await page.goto('/input-password-example-form');
+    await page.goto('/input-password-example-reactive-form');
     const host = getHost(page);
 
     await expect(page.locator('[data-field="touched"]')).toHaveText('false');