+++ packages/components/src/components/input-date/input-date.spec.ts
@@ -1,4 +1,5 @@
 import { vi } from 'vitest';
+import * as implicitSubmitUtils from '../../utils/form/implicitSubmit';
 import { InputDate } from './input-date';
 
 vi.mock('../../utils/dom');
@@ -112,6 +113,17 @@ describe('onInput', () => {
     expect(component.value).toBe(testValue);
   });
 });
+
+describe('onKeyDown', () => {
+  it('should call implicitSubmit with correct parameters', () => {
+    const component = initComponent();
+    const event = new KeyboardEvent('keydown', { key: 'Enter' });
+    const implicitSubmitSpy = vi.spyOn(implicitSubmitUtils, 'implicitSubmit');
+    component['onKeyDown'](event);
+    expect(implicitSubmitSpy).toHaveBeenCalledWith(event, component['internals'], component.host);
+  });
+});
+
 describe('componentDidRender', () => {
   it('should call ElementInternals setValidity()', () => {
     const component = initComponent();