+++ packages/components/src/components/input-email/input-email.spec.ts
@@ -1,5 +1,7 @@
 import { vi } from 'vitest';
 import { InputEmail } from './input-email';
+import * as implicitSubmitUtils from '../../utils/form/implicitSubmit';
+
 
 vi.mock('../../utils/dom');
 
@@ -112,6 +114,17 @@ describe('onInput', () => {
     expect(component.value).toBe(testValue);
   });
 });
+
+describe('onKeyDown', () => {
+  it('should call implicitSubmit with correct parameters', () => {
+    const component = initComponent();
+    const event = new KeyboardEvent('keydown', { key: 'Enter' });
+    const implicitSubmitSpy = vi.spyOn(implicitSubmitUtils, 'implicitSubmit');
+    component['onKeyDown'](event);
+    expect(implicitSubmitSpy).toHaveBeenCalledWith(event, component['internals'], component.host);
+  });
+});
+
 describe('componentDidRender', () => {
   it('should call ElementInternals setValidity()', () => {
     const component = initComponent();