+++ packages/components/src/components/input-search/input-search.tsx
@@ -17,6 +17,7 @@ import {
   FORM_STATES,
   getPrefixedTagNames,
   hasPropValueChanged,
+  implicitSubmit,
   THEMES,
   validateProps,
 } from '../../utils';
@@ -212,6 +213,7 @@ export class InputSearch {
         onInput={this.onInput}
         onChange={this.onChange}
         onBlur={this.onBlur}
+        onKeyDown={this.onKeyDown}
         name={this.name}
         form={this.form}
         type="search"
@@ -249,6 +251,10 @@ export class InputSearch {
       />
     );
   }
+
+  private onKeyDown = (e: KeyboardEvent): void => {
+    implicitSubmit(e, this.internals, this.host);
+  };
 
   private onChange = (e: Event): void => {
     e.stopPropagation();