+++ packages/storefront/src/sitemap.tsx
@@ -175,6 +175,11 @@ export const sitemap: Routes = {
             path: '/developing/angular/demo',
             type: 'TAB',
           },
+          form: {
+            name: 'Form',
+            path: '/developing/angular/form',
+            type: 'TAB',
+          },
           testing: {
             name: 'Testing',
             path: '/developing/angular/testing',
@@ -224,6 +229,11 @@ export const sitemap: Routes = {
             path: '/developing/next-js/getting-started',
             type: 'TAB',
           },
+          form: {
+            name: 'Form',
+            path: '/developing/next-js/form',
+            type: 'TAB',
+          },
           demo: {
             name: 'Demo',
             path: '/developing/next-js/demo',
@@ -254,6 +264,11 @@ export const sitemap: Routes = {
           demo: {
             name: 'Demo',
             path: '/developing/react/demo',
+            type: 'TAB',
+          },
+          form: {
+            name: 'Form',
+            path: '/developing/react/form',
             type: 'TAB',
           },
           testing: {
@@ -288,6 +303,11 @@ export const sitemap: Routes = {
             path: '/developing/react-router/demo',
             type: 'TAB',
           },
+          form: {
+            name: 'Form',
+            path: '/developing/react-router/form',
+            type: 'TAB',
+          },
           faq: {
             name: 'FAQ',
             path: '/developing/react-router/faq',
@@ -315,6 +335,11 @@ export const sitemap: Routes = {
             path: '/developing/vanilla-js/demo',
             type: 'TAB',
           },
+          form: {
+            name: 'Form',
+            path: '/developing/vanilla-js/form',
+            type: 'TAB',
+          },
           advanced: {
             name: 'Advanced',
             path: '/developing/vanilla-js/advanced',
@@ -340,6 +365,11 @@ export const sitemap: Routes = {
           demo: {
             name: 'Demo',
             path: '/developing/vue/demo',
+            type: 'TAB',
+          },
+          form: {
+            name: 'Form',
+            path: '/developing/vue/form',
             type: 'TAB',
           },
           advanced: {